Create Exceptions for NotFound and BadRequest

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.Contracts.Exceptions;
 using HR.LeaveManagement.Application.Contracts.Persistence;
 using HR.LeaveManagement.Domain;
 using MediatR;
@@ -19,7 +20,10 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
         var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
 
         // Verify that record exists
-
+        if(leaveTypeToDelete == null)
+        {
+            throw new NotFoundException(nameof(Domain.LeaveType), request.Id);
+        }
 
         // Delete
         await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
